codeintel: Move datastructures package
