+++ lib/codeintel/lsif/conversion/canonicalize_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 
-	"github.com/sourcegraph/sourcegraph/lib/codeintel/datastructures"
+	"github.com/sourcegraph/sourcegraph/lib/codeintel/lsif/conversion/datastructures"
 )
 
 func TestCanonicalizeDocuments(t *testing.T) {
