+++ lib/codeintel/lsif/conversion/datastructures/compare.go
