+++ lib/codeintel/lsif/conversion/datastructures/disjoint_idset.go
